initial movement of operations code into separate files
